% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtick-package.R
\docType{package}
\name{backtick-package}
\alias{backtick}
\alias{backtick-package}
\title{backtick: Insert Backticks And Backtick Structures}
\description{
Insert a backtick, surround text with backticks, or surround text with backticks to demarcate as inline or chunked R code in an R markdown document. Intended for use as an RStudio Addin.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
