% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backticks.R
\name{bt_backtick}
\alias{bt_backtick}
\title{Insert Backtick}
\usage{
bt_backtick()
}
\description{
Insert a single backtick. In R Markdown file, one backtick will be inserted.
RStudio automatically adds a second backtick when this function is used in
an R script.
}
