% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backticks.R
\name{bt_backticks}
\alias{bt_backticks}
\title{Insert Backticks}
\usage{
bt_backticks()
}
\description{
Surround selected text with backticks. Use it to style strings as code when
rendering an R Markdown file, or to reference variable names that contain
spaces.
}
