% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backticks.R
\name{bt_chunk}
\alias{bt_chunk}
\title{Insert Chunk}
\usage{
bt_chunk()
}
\description{
Place selected text inside an R Markdown chunk. Places triple backticks above
and below the selection, with an 'r' in curly braces after the first set. The
content will be executed on render.
}
