% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backticks.R
\name{bt_inline}
\alias{bt_inline}
\title{Insert Inline R Code}
\usage{
bt_inline()
}
\description{
Surround selected text with backticks and begin with an 'r'. Use this to
demarcate inline R code in R Markdown so that it's executed on render.
}
