% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtick-package.R
\docType{package}
\name{backtick-package}
\alias{backtick}
\alias{backtick-package}
\title{backtick: Insert Backticks And Backtick Structures}
\description{
Insert into an R script a backtick; surround text with backticks; or surround text with backticks to demarcate inline or chunked R code in an R Markdown document. Intended for use as an RStudio Addin.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/co-analysis/backtick}
  \item Report bugs at \url{https://github.com/co-analysis/backtick/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
